+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed clicking "Pause All" and "Resume All" buttons result in crashing.
 - Fixed download button and connections dropdown menu still enabled after download has started resulting in possible launch of invalid downloads.
+- Fixed download button and connections dropdown menu keep disabled if the user entered an invalid url, by improving url validation logic.
 
 ### Changed
 - Replaced the Warning dialogs with native windows dialogs.